Removed IIFE
I wanted to create a way to check if the data in the localStorage is the same with that in the JSON files. Using the initial process won't be scalable when I most to a DB and automate most things. However, if I use *registeredLastUpdate* as a *Date*, I can timestamp the last update and see if localStorage is up to date.