Add convertLocalToModel function to controller
convertLocalToModel function is used to convert LocalStorage real estate parameter to model. It makes use of for-in loop sice it's fast for iterating across Objects. If the localstroage is not found, it logs an Error. The function is only called when we are sure we have the parameter we need in the localStorage (: