Removes destructuring in Classes
As much as I think destructuring will ensure that my data and classes adhere to a certain structure, I think for now it force me to declare some unnecessary parameter and/or create an unnecessary strict and unscalable structure. For instance, I most likely won't use isFeatured parameter in the property class but it's added because of destructuring of my data/object.

I think as my knowledge expands, I will know if this is best.