Create render for home, renderHome
This function is helps create the element we need for home page. FOr now, it basically delcare object paramenters of what is need.

NB: Partners were in FIgma but the object weren't added to the data file. Thanks to MVC coding pattern, we can easily incorporate partners later. Also, I will need to make a different function for object which can be called across page. Angular would have made this easier for me.